moved learn code to dedicated ablities
