+++ module/data/item/talent.mjs
@@ -2,6 +2,7 @@ import AbilityTemplate from "./templates/ability.mjs";
 import ItemDescriptionTemplate from "./templates/item-description.mjs";
 import ED4E from "../../config.mjs";
 import KnackTemplate from "./templates/knack-item.mjs";
+import PromptFactory from "../../applications/global/prompt-factory.mjs";
 
 /**
  * Data model template with information on talent items.
@@ -256,6 +257,26 @@ export default class TalentData extends AbilityTemplate.mixin(
   /** @inheritDoc */
   static async learn( actor, item, createData = {} ) {
     const learnedItem = await super.learn( actor, item, createData );
+
+    let category = null;
+
+    // assign the talent category
+    const promptFactoryItem = PromptFactory.fromDocument( learnedItem );
+    category = await promptFactoryItem.getPrompt( "talentCategory" );
+
+    // assign the level at which the talent was learned
+    const promptFactoryActor = PromptFactory.fromDocument( actor );
+    const disciplineUuid = await promptFactoryActor.getPrompt( "chooseDiscipline" );
+    const discipline = await fromUuid( disciplineUuid );
+    const learnedAt = discipline?.system.level;
+
+    await learnedItem.update( {
+      "system.talentCategory":        category,
+      "system.source.class":          discipline ? discipline.uuid : null,
+      "system.source.atLevel":        learnedAt ? learnedAt : null,
+    } );
+    await learnedItem.system.chooseTier();
+    
     return learnedItem;
   }
